refactor: move DrawingParametersEnum to common, so that importing OSMDOptions doesn't cause import chain to EngravingRules etc
same will be done for ColoringModes
