+++ src/Common/Enums/index.ts
@@ -1,5 +1,6 @@
 // created from 'create-ts-index'
 
+export * from "./DrawingParametersEnum";
 export * from "./FontStyles";
 export * from "./Fonts";
 export * from "./TextAlignment";
