+++ src/MusicalScore/Graphical/DrawingParameters.ts
@@ -1,20 +1,11 @@
 import { EngravingRules } from "./EngravingRules";
 import { PlacementEnum } from "../VoiceData/Expressions/AbstractExpression";
+import { DrawingParametersEnum } from "../../Common/Enums/DrawingParametersEnum";
 
 export enum ColoringModes {
     XML = 0,
     AutoColoring = 1,
     CustomColorSet = 2
-}
-
-export enum DrawingParametersEnum {
-    allon = "allon",
-    compact = "compact",
-    compacttight = "compacttight",
-    default = "default",
-    leadsheet = "leadsheet",
-    preview = "preview",
-    thumbnail = "thumbnail",
 }
 
 /** Internal drawing/rendering parameters and broad modes like compact and thumbnail. Overlap with EngravingRules. */
