+++ src/OpenSheetMusicDisplay/OpenSheetMusicDisplay.ts
@@ -12,7 +12,8 @@ import { Cursor } from "./Cursor";
 import { MXLHelper } from "../Common/FileIO/Mxl";
 import { AJAX } from "./AJAX";
 import log from "loglevel";
-import { DrawingParametersEnum, DrawingParameters, ColoringModes } from "../MusicalScore/Graphical/DrawingParameters";
+import { DrawingParameters, ColoringModes } from "../MusicalScore/Graphical/DrawingParameters";
+import { DrawingParametersEnum } from "../Common/Enums/DrawingParametersEnum";
 import { IOSMDOptions, OSMDOptions, AutoBeamOptions, BackendType, CursorOptions } from "./OSMDOptions";
 import { EngravingRules, PageFormat } from "../MusicalScore/Graphical/EngravingRules";
 import { AbstractExpression } from "../MusicalScore/VoiceData/Expressions/AbstractExpression";
